+++ package.json
@@ -22,7 +22,7 @@
     "@haqq/provider-hot-react-native": "0.0.5",
     "@haqq/provider-ledger-react-native": "0.0.18",
     "@haqq/provider-mnemonic-react-native": "0.0.5",
-    "@haqq/provider-mpc-react-native": "0.0.3",
+    "@haqq/provider-mpc-react-native": "0.0.4",
     "@haqq/provider-web3-utils": "0.0.9",
     "@invertase/react-native-apple-authentication": "^2.2.2",
     "@react-native-async-storage/async-storage": "1.17.11",
